% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_dataset_validation.R
\name{as.frequency.matrix}
\alias{as.frequency.matrix}
\title{Convert confusion matrices and tables to frequency matrices}
\usage{
as.frequency.matrix(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{}
}
\value{
a frequency matrix ormalizing by the column total
}
\description{
Convert confusion matrices and tables to frequency matrices
}
\examples{
table(1:3, 3:1)

# 1 2 3
# 1 0 0 1
# 2 0 1 0
# 3 1 0 0
as.frequency.matrix(table(1:3, 3:1))
#
# 1   2   3
# 1   0   0 100
# 2   0 100   0
# 3 100   0   0
as.frequency.matrix(table(c(3,3,2), 3:1))
#
#     2   3
# 1 100   0
# 2   0 100
# 3   0 100
}
