% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tree_rules.R
\name{as.garnett}
\alias{as.garnett}
\alias{print.garnett.list}
\title{Converts a classification tree to a garnett output}
\usage{
as.garnett(tree, digits = 3, rules_keep = ".*")

\method{print}{garnett.list}(x)
}
\arguments{
\item{tree}{a fitted constparty/party object}

\item{digits}{an integer stating how many decimal places should be conserved}

\item{rules_keep}{a regular expression stating which rules should be kept,
defaults to all rules (".*")}
}
\value{
a complex list that prints as a garnett file
}
\description{
Converts a fitted ctree object to the classifier format supported by garnett
for more information on the specification please visit:
https://cole-trapnell-lab.github.io/garnett/docs/#submitting-a-classifier
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Print a garnett classifier
}}

\examples{
as.garnett(fit_ctree(sctree::small_5050_mix))
# > 0_node_3 	(n = 40)
# expressed below: ADA 2.837, ASNS 1.432
# 
# > 0_node_6 	(n = 8)
# expressed above: ASNS 1.432, ITGA4 0
# expressed below: ADA 2.837
# 
# > 0_node_8 	(n = 135)
# expressed above: ADA 2.837
# expressed below: TSC22D3 2.121
# 
# > 0_node_9 	(n = 18)
# expressed above: ADA 2.837, TSC22D3 2.121
# 
# > 1 	(n = 54)
# expressed above: ASNS 1.432
# expressed below: ADA 2.837, ITGA4 0
}
