% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{autoplot}
\alias{autoplot}
\title{autoplot
wrapper arround ggplot2 autoplot generic}
\usage{
autoplot(object, ...)
}
\arguments{
\item{object}{an object ...}

\item{...}{additional arguments passed to the correct method}
}
\description{
autoplot
wrapper arround ggplot2 autoplot generic
}
