% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_dataset_validation.R
\name{autoplot.table}
\alias{autoplot.table}
\title{Heatmap with ggplot from table objects}
\usage{
autoplot.table(tbl, min_color = NULL, show_number = FALSE)
}
\arguments{
\item{tbl}{a table object}

\item{min_color}{minimum value to show colored, value sunder this will
be shown as gray}

\item{show_number}{logical indicating wether the number should be printed
in the box of the heatmap}
}
\value{
ggpplot object
}
\description{
Heatmap with ggplot from table objects
}
\examples{
autoplot(table(1:10, 1:10), show_number = TRUE)
}
