% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tree_rules.R
\name{fit_ctree}
\alias{fit_ctree}
\title{Fits a classification tree on a Seurat object}
\usage{
fit_ctree(object, genes_use = Seurat::VariableFeatures(object),
  cluster = "ALL", ...)
}
\arguments{
\item{object}{a Seurat object}

\item{genes_use}{a character vector indicating which genes to use in
the classification. currently implemented only for Seurat objects.
(for data frames one can simply subset the input data frame)
defaults to Seurat::VariableFeatures(object)}

\item{cluster}{a cluster name for which the markers will be found}

\item{...}{additional arguments to be passed to partykit::ctree_control}
}
\value{
a ctree fit
}
\description{
Fits a classification tree on a Seurat object
}
\examples{

   fit_ctree(small_9901_mix, c('CCNB1', 'PLK1', 'AURKA'), cluster = 'ALL')
   fit_ctree(small_9901_mix, c('CCNB1', 'PLK1', 'AURKA'), cluster = '0')
   
# 
# Model formula:
# ident ~ CCNB1 + PLK1 + AURKA
# 
# Fitted party:
# [1] root
# |   [2] PLK1 <= 2.31327
# |   |   [3] CCNB1 <= 3.01197
# |   |   |   [4] PLK1 <= 1.75363: clus 0 (n = 202, err = 4.0%)
# |   |   |   [5] PLK1 > 1.75363: clus 0 (n = 45, err = 20.0%)
# |   |   [6] CCNB1 > 3.01197
# |   |   |   [7] PLK1 <= 1.67634: clus 0 (n = 26, err = 38.5%)
# |   |   |   [8] PLK1 > 1.67634: not clus 0 (n = 23, err = 17.4%)
# |   [9] PLK1 > 2.31327
# |   |   [10] AURKA <= 2.02766: not clus 0 (n = 21, err = 47.6%)
# |   |   [11] AURKA > 2.02766: not clus 0 (n = 67, err = 4.5%)
# 
# Number of inner nodes:    5
# Number of terminal nodes: 6
}
