% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_aliases}
\alias{get_aliases}
\title{get_aliases}
\usage{
get_aliases(gene_symbols, db = org.Hs.eg.db::org.Hs.eg.db)
}
\arguments{
\item{gene_symbols}{a character vector of gene symbols. zb: CD1A}

\item{db}{a database that inherits the select property.
defaults to org.Hs.eg.db}
}
\value{
a named list whose names are the provided symbols and the elements are
    the character vectors with the aliases
}
\description{
gets the aliases for a series of gene names
}
\examples{
get_aliases(c("MAPK1", "CD4"))
# $MAPK1
# [1] "ERK"      "ERK-2"    "ERK2"     "ERT1"     "MAPK2"    "P42MAPK"
# [7] "PRKM1"    "PRKM2"    "p38"      "p40"      "p41"      "p41mapk"
# [13] "p42-MAPK" "MAPK1"
#
# $CD4
# [1] "CD4mut" "CD4"
}
