% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tree_rules.R
\name{get_cluster_mapping}
\alias{get_cluster_mapping}
\title{get_cluster_mapping}
\usage{
get_cluster_mapping(tree)
}
\arguments{
\item{tree}{a partykit::ctree fit}
}
\value{
a named vector
}
\description{
returns a named vector, the names being the terminal node id in the tree
and the values being the classification it corresponds to
}
\examples{
my_tree <- fit_ctree(small_5050_mix, c('CKB', 'ADA', 'ASNS', 'PRDX6', 'MZB1'))
get_cluster_mapping(my_tree)

}
#  3 5 6 8 9 
#  0 0 1 0 1 
#  Levels: 0 1
