% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tree_rules.R
\name{get_concensus_rules}
\alias{get_concensus_rules}
\title{Consolidates the rules in a decision tree per cluster}
\usage{
get_concensus_rules(tree)
}
\arguments{
\item{tree}{partykit::ctree fit decision tree}
}
\value{
a nested list
}
\description{
Returns a nested list of the concensus rules to define a cluster, given a
partykit decision tree.
}
\details{
level 1 of the nested list is the group being predicted.
level 2 is either "all" or "majority".
each element is a character vector.
}
\examples{

    iris_tree <- partykit::ctree(Species ~ ., data = iris)
    str(get_concensus_rules(iris_tree))
    
# List of 3
#  $ setosa    :List of 2
#   ..$ all     : chr "Petal.Length <= 1.9"
#   ..$ majority: chr(0) 
#  $ versicolor:List of 2
#   ..$ all     : chr [1:2] "Petal.Length > 1.9" "Petal.Width <= 1.7"
#   ..$ majority: chr "Petal.Length <= 4.8"
#  $ virginica :List of 2
#   ..$ all     : chr [1:2] "Petal.Length > 1.9" "Petal.Width > 1.7"
#   ..$ majority: chr(0) 
#  - attr(*, "class")= chr [1:2] "list" "concensus.rules"
}
