% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_genesymbols}
\alias{get_genesymbols}
\title{get_genesymbols}
\usage{
get_genesymbols(gene_aliases, db = org.Hs.eg.db::org.Hs.eg.db)
}
\arguments{
\item{gene_aliases}{a character vector of gene aliases zb: ERK}

\item{db}{a database that inherits the select property.
defaults to org.Hs.eg.db}
}
\value{
a named list whose names are the provided aliases and the elements
    are the character vectors with the gene names
}
\description{
gets the aliases for a series of common gene names
}
\examples{
get_genesymbols("ERK")
# $ERK
# [1] "EPHB2" "MAPK1"
get_genesymbols(c("SUPERFAKEGENE", "ERK"))
# $SUPERFAKEGENE
# [1] NA
#
# $ERK
# [1] "EPHB2" "MAPK1"
}
\seealso{
get_aliases
}
