% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggheatmap_base}
\alias{ggheatmap_base}
\title{Basic heatmap using ggplot}
\usage{
ggheatmap_base(
  df,
  x = "cluster",
  y = "predicted",
  value = "Freq",
  show_number = FALSE
)
}
\arguments{
\item{df}{data.frame}

\item{x}{column name used in the x axis}

\item{y}{column name used in the y axis}

\item{value}{value of the color to be used}

\item{show_number}{logical indicating wether the value of z should be printed
in the plot.}
}
\value{
a ggplot object
}
\description{
Basic heatmap using ggplot
}
\examples{
my_df <- data.frame(x = 1:2, y = 1:2, z = 1:2)
ggheatmap_base(my_df, x = "x", y = "y", value = "z", show_number = TRUE)
ggheatmap_base(my_df, x = "x", y = "y", value = "z", show_number = FALSE)
}
\keyword{internal}
