% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{include_roxygen_example}
\alias{include_roxygen_example}
\title{include_roxygen_example}
\usage{
include_roxygen_example(example_string)
}
\arguments{
\item{example_string}{the string to be interpreted as R code}
}
\value{
a string that can be interpreted literally as a roxygen comment
}
\description{
Include the output of code in your documentation automatically
}
\details{
This function takes a string that can be interpreted as r code and returns it
formatted as a roxygen exmaple, including tags and output commented out.

It is meant to be used with the evalRd tag as
`#' @evalRd include_roxygen_example("print('a')")`
}
