% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.confusion.matrix}
\alias{is.confusion.matrix}
\title{Check if the element has the structure of a confusion matrix}
\usage{
is.confusion.matrix(x)
}
\arguments{
\item{x}{table object}
}
\value{
logical indicating wether the element matches the description
}
\description{
Slim check for the element to assure it has 2 dimensions,
is a table and whose elements are exclusively integers
}
\examples{
is.confusion.matrix(table(1:3, 3:1, 1:3)) # FALSE
is.confusion.matrix(table(1:3, 3:1)) # TRUE
}
\keyword{internal}
