% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_gene_membrane}
\alias{is_gene_membrane}
\title{Queries gene symbols and returns if they are annotated as membrane-localized}
\usage{
is_gene_membrane(
  gene_symbols,
  db = org.Hs.eg.db::org.Hs.eg.db,
  evidence_codes = c("EXP", "IDA", "IPI", "IMP", "IGI")
)
}
\arguments{
\item{gene_symbols}{a character vector of gene symbols. zb: CD1A}

\item{db}{a database that inherits the select property.
defaults to org.Hs.eg.db}

\item{evidence_codes}{the evidence codes that will be considered for the
annotation. if set to null will return all annotations. defaults to
c("EXP", "IDA", "IPI", "IMP", "IGI")}
}
\value{
logical vector indicating which gene_symbols match the criteria.
}
\description{
returns a logical vector indicating wether there is a GO annotation
corresponding to *integral component of membrane*. the species to be queried
can be changed by using the corresponging database.
}
\details{
Can also filter the annotations by the level of annotation desired.
For additional information on the accepted annotation codes please visit
http://geneontology.org/docs/guide-go-evidence-codes/
}
\examples{
is_gene_membrane(c("CD4", "UPP1"))
# [1]  TRUE FALSE
}
