% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowstyle_plot.R
\name{plot_flowstyle}
\alias{plot_flowstyle}
\title{plot_flowstyle}
\usage{
plot_flowstyle(object, markernames)
}
\arguments{
\item{object}{data, either a Seurat of a data frame object}

\item{markernames}{names of the markers to be used for the plot}
}
\value{
a ggplot grid with the plots
}
\description{
Simulates flow cytometry plots for single cell rna seq data
}
\details{
Note that it imputes a little bit of noise to values of 0 just for the
Sake of visualization
}
\examples{
plot_flowstyle(Seurat::pbmc_small, c("ACRBP", "TSC22D1", "VDAC3"))
}
