% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibodies.R
\name{query_biolegend_antibodies}
\alias{query_biolegend_antibodies}
\title{Query Biolegend for antibodies}
\usage{
query_biolegend_antibodies(search_term)
}
\arguments{
\item{search_term}{The term used to query the webpage}
}
\value{
a length 0 character when no results are found or a character vector with the product names
}
\description{
Queries Biolegend for antibodies and
filters for antibodies for Flow Cytometry
}
\details{
Note that it is just a wrapper arround the web search of the site, so please
be nice on them and dont get your IP banned ...
}
\examples{
> query_biolegend_antibodies("CD11bfakename")
NULL
> head(query_biolegend_antibodies("CD11C"))
[1] "MojoSortâ\\u0084¢ Mouse CD11c Nanobeads" "APC anti-human CD11c Antibody"          "Biotin anti-human CD11c Antibody"
[4] "FITC anti-human CD11c Antibody"         "PE anti-human CD11c Antibody"           "PE/Cy5 anti-human CD11c Antibody"
}
