% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibodies.R
\name{query_sc_antibodies}
\alias{query_sc_antibodies}
\title{Query Santa Cruz for antibodies}
\usage{
query_sc_antibodies(search_term)
}
\arguments{
\item{search_term}{The term used to query the webpage}
}
\value{
NULL when no results are found or a DF with information
}
\description{
Queries Santa Cruz for antibodies and
filters for antibodies for IFC, IHC, and FCM
}
\details{
Note that it is jsut a wrapper arround the web search of the site, so please
be nice on them and dont get your IP banned ...
}
\examples{
> query_sc_antibodies("CD11bfakename")
NULL
> head(query_sc_antibodies("CD11C"))
Product_Name  Cat_num     Epitope    Species                       Method
1 Integrin alpha X Antibody (2Q862) sc-71454 Integrin aX  and human                   IF and FCM
2 Integrin alpha X Antibody (2Q865) sc-71455 Integrin aX      human                   IP and FCM
3   Integrin alpha X Antibody (3.9)  sc-1185 Integrin aX      human               IP, IF and FCM
4 Integrin alpha X Antibody (3H986) sc-71456 Integrin aX      mouse                   IF and FCM
5   Integrin alpha X Antibody (B-6) sc-46676 Integrin aX      human WB, IP, IF, IHC(P) and ELISA
6 Integrin alpha X Antibody (B-Iy6) sc-19989 Integrin aX      human                   IF and FCM
}
