% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_utils.R
\name{ranger_importances.df}
\alias{ranger_importances.df}
\alias{ranger_importances.Seurat}
\alias{FindAllMarkers_ranger.Seurat}
\title{ranger_importances

Calculates ranger-based variable importances for data frames and
Seurat objects}
\usage{
ranger_importances.df(object, cluster = NULL, pval_cutoff = 0.05,
  imp_method = c("janitza", "altmann"), num.trees = 500,
  warn.imp.method = TRUE, identity_col_name = "ident",
  return_what = "signif_importances_ranger", sort_results = TRUE, ...)

ranger_importances.Seurat(object, cluster = NULL, pval_cutoff = 0.05,
  imp_method = c("janitza", "altmann"), num.trees = 500,
  genes_use = Seurat::VariableFeatures(object), warn.imp.method = TRUE,
  ...)

FindAllMarkers_ranger.Seurat(object,
  genes_use = Seurat::VariableFeatures(object), ...)
}
\arguments{
\item{object}{a Seurat or data frame object}

\item{cluster}{a cluster name for which the markers will be found}

\item{pval_cutoff}{p value cutoff for the markers}

\item{imp_method}{importance method, either of "janitza" or "altmann"}

\item{num.trees}{number of trees to be build using ranger}

\item{warn.imp.method}{logical indicating wether warning should be issued
when few negative importances are found to calculate the p.values in
ranger.}

\item{return_what}{a subset of "ranger_fit", "importances_ranger",
"signif_importances_ranger", defaults to signif_importances_ranger}

\item{...}{additional arguments to be passed to ranger}

\item{genes_use}{a character vector indicating which genes to use in
the classification. currently implemented only for Seurat objects.
(for data frames one can simply subset the input data frame)}

\item{return}{ranger_fit, importances_ranger, signif_importances_ranger}
}
\value{
by default returns a data frame with the importances and p values
    but this behavior can be modified by the
}
\description{
ranger_importances
}
\section{Functions}{
\itemize{
\item \code{ranger_importances.Seurat}: Calculate variable importances to calssify a Seurat object

\item \code{FindAllMarkers_ranger.Seurat}: Calculate variable importances
to each cluster in a Seurat object
}}

\examples{
head(ranger_importances.Seurat(Seurat::pbmc_small, cluster = "ALL", warn.imp.method = FALSE))
}
#           importance pvalue     gene
#  S100A8    7.4340993      0   S100A8
#  HLA-DPB1  6.8053140      0 HLA-DPB1
#  S100A9    5.3470476      0   S100A9
#  HLA-DQA1  2.6894632      0 HLA-DQA1
#  GNLY      2.0230294      0     GNLY
#  IGLL5     0.3928885      0    IGLL5
#  PF4       0.3379842      0      PF4
#  AKR1C3    0.3377994      0   AKR1C3
#  PGRMC1    0.2748009      0   PGRMC1
#  SDPR      0.2474737      0     SDPR
