% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_utils.R
\name{ranger_importances}
\alias{ranger_importances}
\alias{FindAllMarkers_ranger.Seurat}
\title{ranger_importances}
\usage{
ranger_importances(x, ...)

FindAllMarkers_ranger.Seurat(
  object,
  genes_use = Seurat::VariableFeatures(object),
  ...
)
}
\arguments{
\item{...}{additional arguments to be passed to ranger}

\item{object}{a Seurat or data frame object}

\item{genes_use}{a character vector indicating which genes to use in
the classification. currently implemented only for Seurat objects.
(for data frames one can simply subset the input data frame)}

\item{cluster}{a cluster name for which the markers will be found}

\item{pval_cutoff}{p value cutoff for the markers}

\item{imp_method}{importance method, either of "janitza" or "altmann"}

\item{num.trees}{number of trees to be build using ranger}

\item{return_what}{a subset of "ranger_fit", "importances_ranger",
"signif_importances_ranger", defaults to signif_importances_ranger}

\item{warn.imp.method}{logical indicating wether warning should be issued
when few negative importances are found to calculate the p.values in
ranger.}

\item{return}{ranger_fit, importances_ranger, signif_importances_ranger}
}
\value{
by default returns a data frame with the importances and p values
    but this behavior can be modified by the
}
\description{
Calculates ranger-based variable importances for data frames and
Seurat objects
}
\section{Functions}{
\itemize{
\item \code{FindAllMarkers_ranger.Seurat}: Calculate variable importances to each cluster in a Seurat object
}}

\examples{

    head(ranger_importances(Seurat::pbmc_small, cluster = 'ALL', warn.imp.method = FALSE))
    
#          importance pvalue     gene
# HLA-DPB1  7.4731067      0 HLA-DPB1
# S100A8    7.1889368      0   S100A8
# S100A9    5.9627401      0   S100A9
# HLA-DQA1  3.1198601      0 HLA-DQA1
# GNLY      1.7252752      0     GNLY
# IGLL5     0.4638471      0    IGLL5
}
