% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_utils.R
\name{RangerDE}
\alias{RangerDE}
\title{RangerDE}
\usage{
RangerDE(data.use, cells.1, cells.2, verbose, ...)
}
\arguments{
\item{data.use}{sparse matrix passed by seurat}

\item{cells.1}{barcode names of the cells asigned as cluster 1}

\item{cells.2}{barcode names of the cells asigned as cluster 2}

\item{verbose}{logical indicating wether the run should be verbose}

\item{...}{additional arguments passed to sctree::ranger_importances.df}
}
\description{
A Helper function that interfaces with Seurat::FindMarkers
}
\keyword{internal}
\examples{
library(Seurat)
    library(sctree)
    head(FindMarkers(object = sctree::small_5050_mix,
    ident.1 = 0, test.use = 'RangerDE', warn.imp.method = FALSE))
#         importance p_val    gene avg_logFC pct.1 pct.2 p_val_adj
# MZB1     1.0278989     0    MZB1  2.004299 0.828 0.227         0
# CD3G     0.9608472     0    CD3G  1.849551 0.744 0.107         0
# CD3D     4.2636666     0    CD3D  1.842246 0.872 0.373         0
# ARHGDIB  2.8168178     0 ARHGDIB  1.787994 0.856 0.320         0
# AIF1     0.2818278     0    AIF1  1.759311 0.756 0.173         0
# FYB      0.7509669     0     FYB  1.750418 0.739 0.107         0
}
