% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{top_n}
\alias{top_n}
\title{top_n}
\usage{
top_n(df, n, wt)
}
\arguments{
\item{df}{a data frame}

\item{n}{the number of falues to be returned}

\item{wt}{the column to be used for the ranking}
}
\value{
a data frame of length (number of rows) n
}
\description{
same as a less robust dplyr::top_n but with standard eval ....
rank gives value of 1 to the minimum
Whenever there is a tie, returns all elements that tied ...
}
\examples{
top_n(iris, 2, "Sepal.Width")
# Sepal.Length Sepal.Width Petal.Length Petal.Width Species
# 16          5.7         4.4          1.5         0.4  setosa
# 34          5.5         4.2          1.4         0.2  setosa
top_n(iris, -2, "Sepal.Width")
# Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
# 61           5.0         2.0          3.5         1.0 versicolor
# 63           6.0         2.2          4.0         1.0 versicolor
# 69           6.2         2.2          4.5         1.5 versicolor
# 120          6.0         2.2          5.0         1.5  virginica
}
