% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tsne_plot}
\alias{tsne_plot}
\title{TSNE plot generation}
\usage{
tsne_plot(object, fraction = 0.1, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{fraction}{fraction of the data to be used for plotting, defaults to 0.1}

\item{...}{Additional arguments to be passed to Seurat::TSNEPlot}
}
\value{
a ggplot object with the tsne plot
}
\description{
slim wrapper arround Seurat::TSNEPlot that overwrites the return option and
by default uses a tenth of the data.
}
\examples{
tsne_plot(Seurat::pbmc_small)
}
