% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.data.frame.Seurat}
\alias{as.data.frame.Seurat}
\title{Gets a data frame from a Seurat object}
\usage{
\method{as.data.frame}{Seurat}(x, genes, fix_names = FALSE)
}
\arguments{
\item{x}{A Seurat object}

\item{genes}{genes to extract to the data.frame}

\item{fix_names}{logical value indicating wether the gene names should be
converted to R-compatible names. defaults to FALSE}
}
\value{
a data frame.
}
\description{
Provided a Seurat object, returs a data frame of the count values, being
the columns each 'gene' and the rows each UMI/cell.
}
\details{
It returns only the genes annotated as variable and the identity column.
}
\examples{
as.data.frame(Seurat::pbmc_small,
    Seurat::VariableFeatures(Seurat::pbmc_small))[1:3,1:3]
#                     LTB EAF2 CD19
# ATGCCAGAACGACT 6.062788    0    0
# CATGGCCTGTGCAT 6.714813    0    0
# GAACCTGATGAACC 7.143118    0    0
}
