% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.data.frame.Seurat}
\alias{as.data.frame.Seurat}
\title{Gets a data frame from a seurat object}
\usage{
\method{as.data.frame}{Seurat}(seurat, genes)
}
\arguments{
\item{seurat}{A seurat object}
}
\value{
a data frame.
}
\description{
Provided a seurat object, returs a data frame of the count values, being
the columns each 'gene' and the rows each UMI/cell.
}
\details{
It returns only the genes annotated as variable and the identity column.
}
\examples{
> as.data.frame.Seurat(pbmc_small, seurat@var.genes)[1:3,1:3]
                   LTB EAF2 CD19
ATGCCAGAACGACT 6.062788    0    0
CATGGCCTGTGCAT 6.714813    0    0
GAACCTGATGAACC 7.143118    0    0
}
