% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.data.frame.Seurat}
\alias{as.data.frame.Seurat}
\title{Gets a data frame from a Seurat object}
\usage{
\method{as.data.frame}{Seurat}(x, genes = Seurat::VariableFeatures(x), fix_names = TRUE, ...)
}
\arguments{
\item{x}{A Seurat object}

\item{genes}{genes to extract to the data.frame}

\item{fix_names}{logical value indicating wether the gene names should be
converted to R-compatible names. defaults to FALSE}

\item{...}{additional arguments passed to `Seurat::FetchData`}
}
\value{
a data frame.
}
\description{
Provided a Seurat object, returs a data frame of the count values, being
the columns each 'gene' and the rows each UMI/cell.
}
\details{
It returns only the genes annotated as variable and the identity column.
}
\examples{
as.data.frame(sctree::small_5050_mix,
    Seurat::VariableFeatures(sctree::small_5050_mix))[1:3,1:3]
#                TNFRSF4 RP3.395M20.12      ID3
# ATCTGACTTGCAGT       0             0 3.424120
# GACCAAACTACTCT       0             0 3.802467
# TGACTTACTTGCAG       0             0 2.204367
}
