% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tree_rules.R
\name{as.garnett}
\alias{as.garnett}
\alias{print.garnett.list}
\title{as.garnett}
\usage{
as.garnett(tree, digits = 3, rules_keep = ".*")

\method{print}{garnett.list}(x)
}
\arguments{
\item{tree}{a fitted constparty/party object}

\item{digits}{an integer stating how many decimal places should be conserved}
}
\value{
a complex list that prints as a garnett file
}
\description{
Converts a fitted ctree object to the classifier format supported by garnett
for more information on the specification please visit:
https://cole-trapnell-lab.github.io/garnett/docs/#submitting-a-classifier
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Print a garnett classifier
}}

\examples{
as.garnett(fit_ctree(Seurat::pbmc_small))
# > 0_node_2 	(n = 52)
# expressed below: S100A8 2.523
# 
# > 0_node_4 	(n = 7)
# expressed above: S100A8 2.523
# expressed below: S100A9 0
# 
# > 1 	(n = 21)
# expressed above: S100A8 2.523, S100A9 0
}
