% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{autoplot.table}
\alias{autoplot.table}
\alias{autoplot.matrix}
\title{autoplot.table}
\usage{
autoplot.table(object, min_color = NULL, show_number = FALSE, ...)

autoplot.matrix(object, ...)
}
\arguments{
\item{object}{a table object}

\item{min_color}{minimum value to show colored, value sunder this will
be shown as gray}

\item{show_number}{logical indicating wether the number should be printed
in the box of the heatmap}

\item{...}{additional arguments passed to downstream methods}
}
\value{
ggpplot object
}
\description{
Heatmap with ggplot from table objects
}
\section{Functions}{
\itemize{
\item \code{autoplot.matrix}: plot a matrix object as a heatmap using ggplot2
}}

\examples{
autoplot(table(1:10, 1:10), show_number = TRUE)
}
