% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_dataset_validation.R
\name{cross_validate}
\alias{cross_validate}
\title{Fits a decision tree in one data set and tests the performance in another}
\usage{
cross_validate(train, test, cluster, genes_use = train@var.genes,
  warn.gene.removal = TRUE, ...)
}
\arguments{
\item{train}{a seurat object to be used for trainning.}

\item{test}{another seurat object to be used for testing.}

\item{cluster}{the cluster whose equivalence needs to be found.}

\item{genes_use}{character vector specifying which genes to use for the
classification, defaults to train@var.genes}

\item{warn.gene.removal}{logical indicating wether to warn the user when
genes are removed because they are missing in one of the datasets.
defults to TRUE}

\item{...}{additional arguments to be passed to ranger_importances.seurat}
}
\value{
a list containing the
    (1) tree fit,
    (2) a summary_table
    (3) the concensus rules of the tree
    (4) ranger_significance_table
    (5) the suggested genes for the gating
}
\description{
Fits a decision tree in one data set and tests the performance in another
}
\examples{
cross_validate(small_5050_mix, small_9901_mix, cluster = "0")
cross_validate(small_5050_mix, small_9901_mix, cluster = "ALL")
}
