% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_feature_plot.R
\name{featureplot_gadget}
\alias{featureplot_gadget}
\title{featureplot_gadget}
\usage{
featureplot_gadget(seurat_object = Seurat::pbmc_small,
  starting_genes = NULL, genes.use = NULL, cache = "./.cache")
}
\arguments{
\item{seurat_object}{a seurat object to explore}

\item{starting_genes}{genes to start with}

\item{genes.use}{genes use for the classification}

\item{cache}{a directory where the caching will occur}
}
\value{
TODO return something ... its a gadget after all ...
}
\description{
Shiny gadget to explore clusters and features for a seurat object.
}
\examples{
\donttest{
featureplot_gadget(seurat_object = small_5050_mix,
                    genes.use = small_5050_mix@var.genes[
                        is_gene_membrane(small_5050_mix@var.genes)],
                    cache = "./.cache")
}
}
