% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_feature_plot.R
\name{featureplot_gadget}
\alias{featureplot_gadget}
\title{featureplot_gadget}
\usage{
featureplot_gadget(seurat_object = Seurat::pbmc_small,
  starting_genes = NULL, cache = "./.cache")
}
\arguments{
\item{seurat_object}{a seurat object to explore}

\item{starting_genes}{genes to start with}
}
\value{
TODO return something ... its a gadget after all ...
}
\description{
Shiny gadget to explore clusters and features for a seurat object.
}
