% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tree_rules.R
\name{get_cluster_mapping}
\alias{get_cluster_mapping}
\title{get_cluster_mapping}
\usage{
get_cluster_mapping(tree)
}
\arguments{
\item{tree}{a partykit::ctree fit}
}
\value{
a named vector
}
\description{
returns a named vector, the names being the terminal node id in the tree
and the values being the classification it corresponds to
}
\examples{
iris_tree <- partykit::ctree(Species ~ ., data = iris)
get_cluster_mapping(iris_tree)

}
#           2          5          6          7 
#      setosa versicolor versicolor  virginica 
#  Levels: setosa versicolor virginica
