% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowstyle_plot.R
\name{plot_flowstyle}
\alias{plot_flowstyle}
\alias{plot_flowstyle.data.frame}
\alias{plot_flowstyle.Seurat}
\title{plot_flowstyle}
\usage{
plot_flowstyle(object, markernames, classif_col = "ident", ...)

\method{plot_flowstyle}{data.frame}(
  object,
  markernames,
  classif_col = "ident",
  warn = TRUE,
  highlight_class = NULL,
  ...
)

\method{plot_flowstyle}{Seurat}(object, markernames, classif_col = "ident", ...)
}
\arguments{
\item{object}{data, either a Seurat of a data frame object}

\item{markernames}{names of the markers to be used for the plot}

\item{classif_col}{name of the classification column to be used for
the grouping, defaults to "ident"}

\item{...}{additional argumetns to be passed to GGally::ggpairs}

\item{warn}{logical indicating if warninf should be shown, defaults to TRUE}

\item{highlight_class}{a character indicating a class if it wants to be highlighted. ie "1"}
}
\value{
a ggplot grid with the plots
}
\description{
Simulates flow cytometry plots for single cell rna seq data
}
\details{
Note that it imputes a little bit of noise to values of 0 just for the
Sake of visualization
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Draw a flowstyle plot from a data.frame

\item \code{Seurat}: Draw a flowstyle plot from a Seurat object
}}

\examples{
plot_flowstyle(Seurat::pbmc_small, c("ACRBP", "TSC22D1", "VDAC3"))
}
