% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tree_rules.R
\name{plot_gates}
\alias{plot_gates}
\title{Plot Decision trees as gates}
\usage{
plot_gates(object, tree, terminal_node)
}
\arguments{
\item{object}{a seurat object that will be used to draw the plots}

\item{tree}{a decision tree that will be used to construct the gates}

\item{terminal_node}{the name of the terminal node that will be used for the gating}
}
\value{
a ggplot object.
}
\description{
Plot Decision trees as gates
}
\examples{
plot_gates(sctree::small_5050_mix, fit_ctree(sctree::small_5050_mix), '5')

}
