% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tree_rules.R
\name{print.concensus.rules}
\alias{print.concensus.rules}
\title{Prints concensus rules}
\usage{
\method{print}{concensus.rules}(x)
}
\arguments{
\item{concensus_rules}{a concensus.rules object output from `sctree::get_concensus_rules`}
}
\value{
returns silently the same object but prints the rules to the console.
}
\description{
Prints in a human readable format a concensus.rules object, usually an output
of `sctree::get_concensus_rules`
}
\examples{
iris_tree <- partykit::ctree(Species ~ ., data = iris)
my_rules <- get_concensus_rules(iris_tree)
print(my_rules)
# Cluster-setosa:
#   all elements:
#     Petal.Length -
# Cluster-versicolor:
#   all elements:
#     Petal.Length +
#     Petal.Width -
#   majority elements:
#     Petal.Length -
# Cluster-virginica:
#   all elements:
#     Petal.Length +
#     Petal.Width +
}
