% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tree_rules.R
\name{print.concensus.rules}
\alias{print.concensus.rules}
\title{Prints concensus rules}
\usage{
\method{print}{concensus.rules}(x)
}
\arguments{
\item{x}{a concensus.rules object output from `sctree::get_concensus_rules`}
}
\value{
returns silently the same object but prints the rules to the console.
}
\description{
Prints in a human readable format a concensus.rules object, usually an output
of `sctree::get_concensus_rules`
}
\seealso{
get_concensus_rules
}
\examples{

    my_tree <- fit_ctree(small_5050_mix, c('CKB', 'ADA', 'ASNS', 'PRDX6', 'MZB1'))
    my_rules <- get_concensus_rules(my_tree)
    print(my_rules)
    
# Cluster-0: 
# 	all elements:
# 		ADA +
# 		ASNS -
# Cluster-1: 
# 	all elements:
# 		ADA -
# 		ASNS +
# 		CKB +
}
