% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibodies.R
\name{query_biocompare_antibodies}
\alias{query_biocompare_antibodies}
\title{Query Biocompare for antibodies}
\usage{
query_biocompare_antibodies(search_term)
}
\arguments{
\item{search_term}{The term used to query the webpage}
}
\value{
NULL when no results are found or a DF with information
}
\description{
Queries Biocompare for antibodies.
https://www.biocompare.com/
}
\details{
Note that it is just a wrapper arround the web search of the site, so please
be nice on them and dont get your IP banned ...
}
\examples{
query_biocompare_antibodies("CD11bfakename")
# NULL
head(query_biocompare_antibodies("CD11C"),3)
}
