% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_utils.R
\name{ranger_importances.df}
\alias{ranger_importances.df}
\alias{ranger_importances.Seurat}
\alias{FindAllMarkers_ranger.Seurat}
\title{ranger_importances}
\usage{
ranger_importances.df(
  object,
  cluster = NULL,
  pval_cutoff = 0.05,
  imp_method = c("janitza", "altmann"),
  num.trees = 500,
  warn.imp.method = TRUE,
  identity_col_name = "ident",
  return_what = "signif_importances_ranger",
  sort_results = TRUE,
  ...
)

ranger_importances.Seurat(
  object,
  cluster = NULL,
  pval_cutoff = 0.05,
  imp_method = c("janitza", "altmann"),
  num.trees = 500,
  genes_use = Seurat::VariableFeatures(object),
  warn.imp.method = TRUE,
  ...
)

FindAllMarkers_ranger.Seurat(
  object,
  genes_use = Seurat::VariableFeatures(object),
  ...
)
}
\arguments{
\item{object}{a Seurat or data frame object}

\item{cluster}{a cluster name for which the markers will be found}

\item{pval_cutoff}{p value cutoff for the markers}

\item{imp_method}{importance method, either of "janitza" or "altmann"}

\item{num.trees}{number of trees to be build using ranger}

\item{warn.imp.method}{logical indicating wether warning should be issued
when few negative importances are found to calculate the p.values in
ranger.}

\item{return_what}{a subset of "ranger_fit", "importances_ranger",
"signif_importances_ranger", defaults to signif_importances_ranger}

\item{...}{additional arguments to be passed to ranger}

\item{genes_use}{a character vector indicating which genes to use in
the classification. currently implemented only for Seurat objects.
(for data frames one can simply subset the input data frame)}

\item{return}{ranger_fit, importances_ranger, signif_importances_ranger}
}
\value{
by default returns a data frame with the importances and p values
    but this behavior can be modified by the
}
\description{
Calculates ranger-based variable importances for data frames and
Seurat objects
}
\section{Functions}{
\itemize{
\item \code{ranger_importances.Seurat}: Calculate variable importances to calssify a Seurat object

\item \code{FindAllMarkers_ranger.Seurat}: Calculate variable importances to each cluster in a Seurat object
}}

\examples{

    head(ranger_importances.Seurat(Seurat::pbmc_small, cluster = 'ALL', warn.imp.method = FALSE))
    
#          importance pvalue     gene
# HLA-DPB1  7.7575463      0 HLA-DPB1
# S100A8    7.0191797      0   S100A8
# S100A9    5.8566047      0   S100A9
# HLA-DQA1  3.2066588      0 HLA-DQA1
# GNLY      1.3493814      0     GNLY
# VDAC3     0.5323662      0    VDAC3
}
