% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_utils.R
\name{ranger_importances.df}
\alias{ranger_importances.df}
\alias{ranger_importances.seurat}
\title{ranger_importances

Calculates ranger-based variable importances for data frames and
seurat objects}
\usage{
ranger_importances.df(object, cluster = NULL, pval_cutoff = 0.05,
  imp_method = c("janitza", "altmann"), num.trees = 500,
  warn.imp.method = TRUE, ...)

ranger_importances.seurat(object, cluster = NULL, pval_cutoff = 0.05,
  imp_method = c("janitza", "altmann"), num.trees = 500,
  genes_use = object@var.genes, warn.imp.method = TRUE, ...)
}
\arguments{
\item{object}{a seurat or data frame object}

\item{cluster}{a cluster name for which the markers will be found}

\item{pval_cutoff}{p value cutoff for the markers}

\item{imp_method}{importance method, either of "janitza" or "altmann"}

\item{num.trees}{number of trees to be build using ranger}

\item{warn.imp.method}{logical indicating wether warning should be}

\item{...}{additional arguments to be passed to ranger}

\item{genes_use}{a character vector indicating which genes to use in
the classification. currently implemented only for seurat objects.
(for data frames one can simply subset the input data frame)}
}
\value{
list with 3 elements ranger_fit, importances_ranger, signif_importances_ranger
}
\description{
ranger_importances
}
\section{Functions}{
\itemize{
\item \code{ranger_importances.seurat}: Calculate variable importances to calssify a seurat object
}}

\examples{
summary(ranger_importances.seurat(Seurat::pbmc_small, cluster = "ALL"))
# Length Class      Mode
# ranger_fit                15     ranger     list
# importances_ranger        62     -none-     numeric
# signif_importances_ranger  3     data.frame list
#' summary(ranger_importances.seurat(Seurat::pbmc_small, cluster = "0"))
}
