% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_utils.R
\name{RangerDE}
\alias{RangerDE}
\title{RangerDE}
\usage{
RangerDE(data.use, cells.1, cells.2, verbose, ...)
}
\arguments{
\item{data.use}{sparse matrix passed by seurat}

\item{cells.1}{barcode names of the cells asigned as cluster 1}

\item{cells.2}{barcode names of the cells asigned as cluster 2}

\item{verbose}{logical indicating wether the run should be verbose}

\item{...}{additional arguments passed to sctree::ranger_importances.df}
}
\description{
A Helper function that interfaces with Seurat::FindMarkers
}
\examples{
library(Seurat)
    library(sctree)
    head(FindMarkers(object = Seurat::pbmc_small,
    ident.1 = 0, test.use = 'RangerDE', warn.imp.method = FALSE))
#        importance p_val   gene avg_logFC pct.1 pct.2 p_val_adj
# GNLY   0.42979254     0   GNLY  4.784185 0.444 0.068         0
# PGRMC1 0.09705335     0 PGRMC1  3.264396 0.222 0.000         0
# GZMB   0.10398463     0   GZMB  2.994887 0.333 0.091         0
# LAMP1  0.17364336     0  LAMP1  2.658571 0.444 0.068         0
# CD7    0.36348563     0    CD7  2.555973 0.528 0.091         0
# NCOA4  0.10846263     0  NCOA4  2.304732 0.306 0.159         0
}
