% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_utils.R
\name{RangerDE}
\alias{RangerDE}
\title{RangerDE}
\usage{
RangerDE(data.use, cells.1, cells.2, verbose, ...)
}
\arguments{
\item{data.use}{sparse matrix passed by seurat}

\item{cells.1}{barcode names of the cells asigned as cluster 1}

\item{cells.2}{barcode names of the cells asigned as cluster 2}

\item{verbose}{logical indicating wether the run should be verbose}

\item{...}{additional arguments passed to sctree::ranger_importances.df}
}
\description{
A Helper function that interfaces with Seurat::FindMarkers
}
\keyword{internal}
\examples{
library(Seurat)
    library(sctree)
    head(FindMarkers(object = Seurat::pbmc_small,
    ident.1 = 0, test.use = 'RangerDE', warn.imp.method = FALSE))
#        importance p_val   gene  avg_logFC pct.1 pct.2 p_val_adj
# GNLY    0.4337917     0   GNLY  4.7841852 0.444 0.068         0
# CD7     0.3529089     0    CD7  2.5559735 0.528 0.091         0
# IL7R    0.1816056     0   IL7R  2.0708996 0.389 0.295         0
# NOSIP   0.2407237     0  NOSIP  2.0604084 0.444 0.250         0
# CCL5    0.5225574     0   CCL5  1.4914821 0.667 0.273         0
# LGALS1  0.3361842     0 LGALS1 -0.9457555 0.500 0.727         0
}
